Promociones/requirements.txt
streamlit==1.42.2
snowflake-connector-python==2.7.12
pandas==1.4.4
pyarrow==8.0.0
numpy==1.23.4
openpyxl==3.0.10